(Backend) More refactoring - Set target to 13.0:11.0 - Fix compile errors on CFNotificationCenterPost - Move libappearancecell.h to public
