+++ AppearanceSelectionTableCell.m
@@ -94,7 +94,7 @@ NSUserDefaults *userDefaults;
     [userDefaults setObject:[NSNumber numberWithInt:self.type] forKey:key];
     [userDefaults synchronize];
 
-    CFNotificationCenterPostNotification(CFNotificationCenterGetDarwinNotifyCenter(), CFSTR(postNotification), NULL, NULL, YES);
+    CFNotificationCenterPostNotification(CFNotificationCenterGetDarwinNotifyCenter(), (__bridge CFStringRef)postNotification, NULL, NULL, YES);
 
     [self.hostController updateForType:self.type];
 }
